refactor(be): extract QuestProgressRecorder from AiCheckService
Reduces AiCheckService constructor from 15 to 14 params by extracting
saveProgress/saveHistory into a dedicated QuestProgressRecorder component.

via [Happy](https://happy.engineering)
